+++ include/core/color/sse3/Colorf.h
@@ -1,7 +1,9 @@
 #pragma once
+
+#include "core/SimdFeatures.h"
 #include <algorithm>
 
-#ifdef __SSE__
+#if RAYTRACER_SIMD_SSE
 
 #include "core/DivisionByZeroException.h"
 #include "core/InequalityOperator.h"